pyproject.toml
[project]
name = "agentic-research-platform"
version = "0.1.0"
description = "Agentic AI + RAG system"
requires-python = ">=3.10"

dependencies = [
    "requests>=2.32.0",
    "python-dotenv>=1.0.1",
    "tenacity>=8.5.0",
    "pydantic>=2.8.2",
    "tavily-python>=0.5.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.2.2",
    "pandas>=2.2.2",
    "tqdm",
    "rich>=13.7.1",
    "orjson>=3.10.6",
    "aiohttp",
    "tiktoken",
    "ipykernel",
    "streamlit>=1.36.0",
    "openai>=1.40.0",
    "sentence-transformers>=3.0.1",
    "numpy>=1.26.4",
]
